+++ src/index.ts
@@ -1,20 +1,15 @@
 import {
   CarterPayload,
   CarterInteraction,
-  CarterConversationEntry,
   CarterSkill,
-  CarterSkillOptions,
-  CarterSkillAction,
+  buildInteraction,
   CarterSkillInstance,
   CarterData,
-  CarterOpenerData,
-  CarterOpenerInteraction,
-  CarterPersonaliseInteraction,
   Logger,
-  isCarterSkillAction,
-  isCarterSkillOptions,
+  isCarterSkill,
   CarterOpenerPayload,
   CarterPersonalisePayload,
+  ForcedBehaviour,
 } from './types';
 import { v1 as uuidv1 } from 'uuid';
 import { DateTime } from 'luxon';
@@ -32,7 +27,7 @@ export const URLS = {
  */
 class Carter {
   apiKey: string;
-  history: CarterConversationEntry[] = [];
+  history: CarterInteraction[] = [];
   skills: CarterSkill[] = [];
   logger: Logger = new logging.NoOpLogger();
   speakDefault: boolean = false;
@@ -58,9 +53,9 @@ class Carter {
   // INTERACTION
   // ========================
 
-  /**
-   * Say something to carter, takes a query parameter containing your message, as well as an options object (optional). Will create a uuid automatically if one is not provided.
-   */
+  // --------------------------------
+  // SAY
+  // --------------------------------
   async say(text: string, playerId?: string, speak?: boolean): Promise<CarterInteraction> {
     if (!text || typeof text !== 'string') {
       throw Error(`Carter.say() requires a string as the first parameter. Received: ${text}.`);
@@ -85,8 +80,10 @@ class Carter {
     const start = now();
     const triggeredSkills: CarterSkillInstance[] = [];
     const executedSkills: CarterSkillInstance[] = [];
-    let data: CarterData;
+    let response: Response | null = null;
+    let data: CarterData | null = null;
     let interaction: CarterInteraction;
+    let errorMessage: string | null = null;
 
     const payload: CarterPayload = {
       key: this.apiKey,
@@ -94,105 +91,80 @@ class Carter {
       playerId,
       speak,
     };
-    const response = await fetch(URLS.say, {
-      method: 'POST',
-      body: JSON.stringify(payload),
-      headers: { 'Content-Type': 'application/json' },
-    });
-
-    if (response.ok) {
+    try {
+      response = await fetch(URLS.say, {
+        method: 'POST',
+        body: JSON.stringify(payload),
+        headers: { 'Content-Type': 'application/json' },
+      });
+    } catch (e) {
+      this.logger.warn(`Carter.say() failed to fetch.`, { interactionID });
+      errorMessage = (e as Error).message;
+      response = null
+    }
+    if (response) {
       try {
         data = (await response.json()) as CarterData;
       } catch (e) {
-        this.logger.error(`Carter.say() failed to parse response as JSON.`, { interactionID });
-        return {
-          id: interactionID,
-          data: undefined,
-          ok: response.ok,
-          statusCode: response.status,
-          statusMessage: response.statusText,
-          payload,
-          triggeredSkills: [],
-          executedSkills: [],
-          timeTaken: Math.round(now() - start),
-
-          response_text: undefined,
-          response_audio: undefined,
-        };
+        data = null
+        this.logger.warn(`Carter.say() failed to parse response as JSON.`, { interactionID });
       }
-    } else {
-      this.logger.error(`Carter.say() received a ${response.status} response. Status text: ${response.statusText}.`, {
-        interactionID,
-      });
-      return {
-        id: interactionID,
-        data: undefined,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        payload,
-        triggeredSkills: [],
-        executedSkills: [],
-        timeTaken: Math.round(now() - start),
-
-        response_text: undefined,
-        response_audio: undefined,
-      };
     }
-
-    for (const behaviour of data.forced_behaviours) {
-      this.logger.debug(`Carter.say() found a forced behaviour: ${behaviour.name}. Checking for registered skill.`, {
-        interactionID,
-      });
-      const skill = this.findSkill(behaviour.name);
-      let newOutput;
-      if (skill) {
-        const skillInstance = new CarterSkillInstance(skill, data.output.text);
-        if (skill.options.auto) {
-          this.logger.debug(
-            `Carter.say() found a registered skill for ${behaviour.name} with auto set to true. Executing and adding to triggered skills.`,
-            { interactionID },
-          );
-          newOutput = await skillInstance.execute();
-          if (newOutput) {
-            data.output.text = newOutput;
+    if (response && data) {
+      for (const behaviour of data.forced_behaviours as ForcedBehaviour[]) {
+        this.logger.debug(`Carter.say() found a forced behaviour: ${behaviour.name}. Checking for registered skill.`, {
+          interactionID,
+        });
+        const skill = this.findSkill(behaviour.name);
+        let newOutput;
+        if (skill) {
+          const skillInstance = new CarterSkillInstance(skill, data.output.text);
+          if (skill.auto) {
+            this.logger.debug(
+              `Carter.say() found a registered skill for ${behaviour.name} with auto set to true. Executing and adding to triggered skills.`,
+              { interactionID },
+            );
+            try {
+              newOutput = await skillInstance.execute();
+            } catch (e) {
+              this.logger.warn(`Carter.say() failed to execute skill.`, { interactionID });
+              skillInstance.errors.push((e as Error).message);
+            }
+            if (newOutput) {
+              data.output.text = newOutput;
+            }
+            executedSkills.push(skillInstance);
+          } else {
+            this.logger.debug(
+              `Carter.say() found a registered skill for ${behaviour.name} with auto set to false. Adding to triggered skills.`,
+              { interactionID },
+            );
+            triggeredSkills.push(skillInstance);
           }
-          executedSkills.push(skillInstance);
-        } else {
-          this.logger.debug(
-            `Carter.say() found a registered skill for ${behaviour.name} with auto set to false. Adding to triggered skills.`,
-            { interactionID },
-          );
-          triggeredSkills.push(skillInstance);
         }
       }
     }
 
-    interaction = {
+    interaction = await buildInteraction({
       id: interactionID,
-      data,
-      ok: response.ok,
-      statusCode: response.status,
-      statusMessage: response.statusText,
+      type: "say",
+      response,
+      carterData: data,
       payload,
+      start,
       triggeredSkills,
       executedSkills,
-      timeTaken: Math.round(now() - start),
-
-      response_text: data.output.text,
-      response_audio: data.output.audio,
-    };
-    const newConversationEntry: CarterConversationEntry = {
-      isoTimestamp: DateTime.now().toISO(),
-      interaction,
-    };
-    this.history.unshift(newConversationEntry);
+      errorMessage: null
+    })
+    this.history.unshift(interaction);
     this.logger.debug(`Carter.say() finished.`, { interactionID });
     return interaction;
   }
 
-  // Start a conversation with Carter
-  async opener(playerId?: string, speak?: boolean | undefined): Promise<CarterOpenerInteraction> {
+  // ---------------------------------
+  // OPENER
+  // ---------------------------------
+  async opener(playerId?: string, speak?: boolean | undefined): Promise<CarterInteraction> {
     if (playerId && typeof playerId !== 'string') {
       throw Error(`Carter.opener() requires a string as the first parameter. Received: ${playerId}.`);
     }
@@ -210,73 +182,52 @@ class Carter {
     const interactionID = uuidv1();
     this.logger.debug(`Carter.opener() called with playerId: ${playerId}.`, { interactionID });
     const start = now();
-    let data: CarterOpenerData;
+    let data: CarterData | null = null;
+    let errorMessage: string | null = null;
+    let response: Response | null = null;
     const payload: CarterOpenerPayload = {
       key: this.apiKey,
       playerId,
       speak,
     };
-    const response = await fetch(URLS.opener, {
-      method: 'POST',
-      body: JSON.stringify(payload),
-      headers: { 'Content-Type': 'application/json' },
-    });
-
-    if (!response.ok) {
-      this.logger.error(
-        `Carter.opener() received a ${response.status} response. Status text: ${response.statusText}.`,
-        { interactionID },
-      );
-      return {
-        id: interactionID,
-        data: undefined,
-        payload,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        timeTaken: Math.round(now() - start),
-
-        response_text: undefined,
-        response_audio: undefined,
-      };
-    }
-
     try {
-      data = (await response.json()) as CarterOpenerData;
-    } catch (e) {
-      this.logger.error(`Carter.opener() failed to parse response as JSON.`, { interactionID });
-      return {
-        id: interactionID,
-        data: undefined,
-        payload,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        timeTaken: Math.round(now() - start),
-
-        response_text: undefined,
-        response_audio: undefined,
-      };
+      response = await fetch(URLS.opener, {
+        method: 'POST',
+        body: JSON.stringify(payload),
+        headers: { 'Content-Type': 'application/json' },
+      });
+    } catch(e) {
+      this.logger.warn(`Carter.opener() failed to fetch.`, { interactionID });
+      errorMessage = (e as Error).message;
+    }
+    if (response) {
+      try {
+        data = await response.json();
+      } catch (e) {
+        this.logger.warn(`Carter.opener() failed to parse response as JSON.`, { interactionID });
+      }
     }
 
-    const interaction: CarterOpenerInteraction = {
+    const interaction = await buildInteraction({
       id: interactionID,
-      data,
+      type: "say",
+      response,
+      carterData: data,
       payload,
-      ok: response.ok,
-      statusCode: response.status,
-      statusMessage: response.statusText,
-      timeTaken: Math.round(now() - start),
+      start,
+      triggeredSkills: null,
+      executedSkills: null,
+      errorMessage
+    })
 
-      response_text: data.sentence,
-      response_audio: undefined,
-    };
     this.logger.debug(`Carter.opener() finished.`, { interactionID });
     return interaction;
   }
 
-  // Personalise text with Carter
-  async personalise(text: string, speak?: boolean | undefined): Promise<CarterPersonaliseInteraction> {
+  // ---------------------------------
+  // PERSONALISE
+  // ---------------------------------
+  async personalise(text: string, speak?: boolean | undefined): Promise<CarterInteraction> {
     if (!text || typeof text !== 'string') {
       throw Error(`Carter.personalise() requires a string as the first parameter. Received: ${text}.`);
     }
@@ -287,97 +238,69 @@ class Carter {
     } else {
       speak = this.speakDefault as boolean;
     }
-
     const interactionID = uuidv1();
     this.logger.debug(`Carter.personalise() called with text: ${text}.`, { interactionID });
     const start = now();
+    let data: CarterData | null = null;
+    let response: Response | null = null;
+    let errorMessage: string | null = null;
     const payload: CarterPersonalisePayload = {
       key: this.apiKey,
       text,
       speak,
     };
-    const response = await fetch(URLS.personalise, {
-      method: 'POST',
-      body: JSON.stringify(payload),
-      headers: { 'Content-Type': 'application/json' },
-    });
-
-    if (!response.ok) {
-      this.logger.error(
-        `Carter.personalise() received a ${response.status} response. Status text: ${response.statusText}.`,
-        { interactionID },
-      );
-      return {
-        id: interactionID,
-        data: undefined,
-        payload,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        timeTaken: Math.round(now() - start),
-
-        response_text: undefined,
-        response_audio: undefined,
-      };
-    }
 
     try {
-      const data = (await response.json()) as CarterData;
-      this.logger.debug(`Carter.personalise() finished.`, { interactionID });
-      return {
-        id: interactionID,
-        data,
-        payload,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        timeTaken: Math.round(Math.round(now() - start)),
-
-        response_text: data.output.text,
-        response_audio: undefined,
-      };
-    } catch (e) {
-      this.logger.error(`Carter.personalise() failed to parse response as JSON.`, { interactionID });
-      return {
-        id: interactionID,
-        data: undefined,
-        payload,
-        ok: response.ok,
-        statusCode: response.status,
-        statusMessage: response.statusText,
-        timeTaken: Math.round(Math.round(now() - start)),
-        response_text: undefined,
-        response_audio: undefined,
-      };
+      response = await fetch(URLS.personalise, {
+        method: 'POST',
+        body: JSON.stringify(payload),
+        headers: { 'Content-Type': 'application/json' },
+      });
+    } catch (error) {
+      this.logger.warn(`Carter.personalise() failed to fetch.`, { interactionID });
+      errorMessage = (error as Error).message;
     }
+    if (response) {
+      try {
+        data = await response.json();
+      }
+      catch (e) {
+        this.logger.warn(`Carter.personalise() failed to parse response as JSON.`, { interactionID });
+        errorMessage = (e as Error).message;
+      }
+    }
+
+    const interaction = await buildInteraction({
+      id: interactionID,
+      type: "personalise",
+      response,
+      carterData: data,
+      payload,
+      start,
+      triggeredSkills: null,
+      executedSkills: null,
+      errorMessage
+    })
+
+    this.logger.debug(`Carter.personalise() finished.`, { interactionID });
+    return interaction;
   }
 
   // SKILLS
   // =====================
 
-  registerSkill(name: string, action: CarterSkillAction, options?: CarterSkillOptions): CarterSkill | undefined {
-    if (!name || typeof name !== 'string') {
-      throw Error(`Carter.registerSkill() requires a string as the first parameter. Received: ${name}.`);
-    }
-    if (!action || !isCarterSkillAction(action)) {
-      throw Error(
-        `Carter.registerSkill() requires a function as the second parameter. Received: ${action}. This function is called with one parameter, the carter response text`,
-      );
-    }
-    if (options && !isCarterSkillOptions(options)) {
-      throw Error(
-        `Carter.registerSkill() requires valid CarterSkillOptions as the third parameter. Received: ${options}.`,
-      );
+  registerSkill(skill: CarterSkill): CarterSkill | undefined {
+    if (!isCarterSkill(skill)) {
+      throw new Error(`Incorrect skill format.`);
     }
 
-    this.logger.debug(`Carter.registerSkill() called with name: ${name}.`);
-    const existing = this.findSkill(name);
+    this.logger.debug(`Carter.registerSkill() called with name: ${skill.name}.`);
+    const existing = this.findSkill(skill.name);
     if (existing) {
-      throw new Error(`A skill with the name ${name} already exists.`);
+      throw new Error(`A skill with the name ${skill.name} already exists.`);
     } else {
-      const skill: CarterSkill = { name, action, options: options ? options : {} };
       this.skills.push(skill);
-      this.logger.debug(`Carter.registerSkill() finished. Registered skill with name: ${name}.`);
+      this.logger.debug(`Carter.registerSkill() finished. Registered skill with name: ${skill.name}.`);
       return skill;
     }
   }
@@ -397,7 +320,7 @@ class Carter {
   /**
    * Returns the latest interaction in the conversation history.
    */
-  latest(): CarterConversationEntry | undefined {
+  latest(): CarterInteraction | undefined {
     if (this.history.length > 0) {
       return this.history[0];
     } else {
@@ -407,7 +330,7 @@ class Carter {
 
   /** Returns the response time of your last interaction with Carter */
   lastResponseTime(): number | undefined {
-    return this.latest()?.interaction.timeTaken;
+    return this.latest()?.timeTaken;
   }
 
   /**
@@ -425,8 +348,8 @@ class Carter {
       const lowerBound = DateTime.now().minus({ minutes });
       for (const entry of this.history) {
         const dt = DateTime.fromISO(entry.isoTimestamp);
-        if (dt > lowerBound && entry.interaction.timeTaken) {
-          times.push(entry.interaction.timeTaken);
+        if (dt > lowerBound && entry.timeTaken) {
+          times.push(entry.timeTaken);
         }
       }
       if (times.length > 0) {
@@ -441,8 +364,8 @@ class Carter {
     } else {
       for (const entry of this.history) {
         const dt = DateTime.fromISO(entry.isoTimestamp);
-        if (entry.interaction.timeTaken) {
-          times.push(entry.interaction.timeTaken);
+        if (entry.timeTaken) {
+          times.push(entry.timeTaken);
         }
       }
       if (times.length > 0) {