+++ src/testing/__tests__/logging.jest.ts
@@ -1,9 +1,13 @@
+import { consoleLogger } from "../../logger";
+import Carter from "../../index";
+import dotenv from 'dotenv';
+dotenv.config()
 // TEST ALL NO OP LOGGER FUNCTIONS
 
 // TEST VALIDATELOGGER
 
 describe('logging', () => {
-  test('placeholder', () => {
-    expect(true).toBe(true);
+  test('passing a logger throws no errors', () => {
+    expect(() => {const carter = new Carter(process.env.CARTER_API_KEY as string, new consoleLogger())}).not.toThrowError();
   });
 });